Use a slice for InstMap instead of std.HashMap
The `sema.inst_map` datastructure is very often accessed. All
instructions that reference the result of other instructions does a
lookup into this field. Because of thisi, a significant amount of time,
is spent in `std.HashMap.get`.

This commit replaces the `HashMap` with a simpler data structure that
uses the zir indexes to index into a slice for the result. See the data
structure doc comment for more info.

Performance results:

  > $(which time) -v perf stat -r 10 stage2-release/bin/zig build-exe \
        ../zig2/src/main.zig --pkg-begin build_options options.zig \
        --pkg-end -lc -fno-emit-bin

                master           new-inst-map
  instructions  15,565,458,749   13,987,614,034  ~10.1% less
      branches   2,547,101,394    2,255,015,269  ~11.5% less
          time        2.40819s         2.19619s   ~8.8% less
                                                 (~9.7% faster)