RunStep: allow emulation to run binaries
When a user of build.zig crosscompiles for a target
that cannot be ran on the host platform, we check if any of
the emulation options are enabled (such as `enable_wasmtime`).
If enabled, we run the compiled binary on the emulation platform/runtime
instead, or provide an error when no emulation can be done.
e.g. no option was enabled.