std.crypto: Add segmented API to Ed25519 signing.
std.crypto.sign.Ed25519.sign previously required the message to be
passed as a single continuous byte slice. This adds a second function
std.crypto.sign.Ed25519.signSegmented which permits the message to be a
slice of byte slices. The original sign function simply wraps the
msg argument in a single item tuple and passes a slice of that to
signSegmented.