copyFile: use FICLONE when possible
Linux supports `ioctl_ficlone(2)`: this is a CoW mechanism to copy a
file. Benefits over `copy_file_range`:
- it's latency does not depend on the file size.
- blocks will be reused, so when the file is copied this way, it will
  not take extra space (until either file is modified).

This is different from a hardlink, because the inode is different, thus
these two files are semantically (from a user's perspective) different.

I suggest making this a default, since, if you do `cp a b` on Linux
since coreutils 9.0, it will reflink by default[1].

As of writing it is supported on btrfs, cifs, nfs, ocfs2, overlayfs and
xfs; for this reason I highly encourage using one of those file systems.
:)

[1]: https://lists.gnu.org/archive/html/info-gnu/2021-09/msg00010.html