copy_file_range: use FICLONERANGE when possible
This is a follow-up from #12476: instead of adding a new abstraction to
clone files, start with `copy_file_range`. They use the same mechanism
in the kernel.

Benefits of FICLONERANGE vs `copy_file_range(2)`:
- O(1).
- CoW: so if files are not modified, it will not take additional space.

However, it is more restricted than copy_file_range: source and
destination must be on the same partition, and only some file systems
implement this. As of Linux 5.19 those are btrfs, cifs, nfs, ocfs2,
overlayfs and xfs[1].

[1]: https://elixir.bootlin.com/linux/v5.19/A/ident/remap_file_range