coff: specify default base path for relative source paths in pdb
PDB requires all file paths to be fully resolved, and it is really
the linker's responsibility to canonicalize any path extracted from
the CodeView in the object file. However, LLD-link has some very
questionable defaults, and in particular, it purposely bakes in path
separator of the host system it was built on rather the target's, or
just throw an error. Thankfully, they have left a backdoor we can
use via `-PDBSOURCEPATH` flag.