std.os.linux: add ioctlFunc
Adds ioctlFunc which returns a wrapper function whose parameter type
is tied to the ioctl number/definition.  This removes the need for using
`@ptrToInt` for the ioctl argument.  Instead, the argument will be
properly declared as a pointer and is guaranteed to have the correct
size and IO direction (aka mutability).  These two guarantees prevent
out-of-bound memory access caused from a mismatch in what the application
passes to the ioctl argument.